switch to use Output Feedback Split mode
https://github.com/waterret/RngState-OFS
